requirements.txt
streamlit
requests
python-dotenv
duckduckgo-search
